Refactor VehicleServiceTableRow to not use <table> element
Also change the font for hours in this component to monospace, to fix a minor visual bug.
We can consider later to import monospaced arial font and fix this with that.

Resolves HSLdevcom/jore4#1197
