+++ ui/src/components/timetables/vehicle-schedule-details/vehicle-service-table/index.ts
@@ -1,1 +1,2 @@
+export * from './VehicleServiceRow';
 export * from './VehicleServiceTable';
